requirements.txt
bcrypt==3.2.0
cffi==1.14.4
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.22
protobuf==3.14.0
pycparser==2.20
PyJWT==1.7.1
six==1.15.0
Werkzeug==1.0.1